pyproject.toml
[project]
name = "videospark-backend"
version = "0.1.0"
description = "FastAPI + single-worker backend scaffold for the VideoSpark project."
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "httpx>=0.28.0",
  "google-genai>=1.0.0",
  "pydantic>=2.10.0",
  "pydantic-settings>=2.7.0",
  "PyJWT[crypto]>=2.10.0",
  "structlog>=24.4.0",
  "supabase>=2.13.0",
  "uvicorn[standard]>=0.34.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.25.0",
  "ruff>=0.9.0",
]

[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
markers = [
  "live_elevenlabs: calls the real ElevenLabs TTS API and writes narration artifacts",
  "live_veo: calls the real Gemini Veo API and downloads a generated clip",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
